Module Ast_403.Docstrings

Docstrings

type docstring

Documentation comments

val docstring : string -> Location.t -> docstring

Create a docstring

val docstring_body : docstring -> string

Get the text of a docstring

val docstring_loc : docstring -> Location.t

Get the location of a docstring

Items

type docs = {

docs_pre : docstring option;

docs_post : docstring option;

}

val empty_docs : docs

val docs_attr : docstring -> Parsetree.attribute

val add_docs_attrs : docs -> Parsetree.attributes -> Parsetree.attributes

Convert item documentation to attributes and add them to an attribute list

Fields and constructors

type info = docstring option

val empty_info : info

val info_attr : docstring -> Parsetree.attribute

val add_info_attrs : info -> Parsetree.attributes -> Parsetree.attributes

Convert field info to attributes and add them to an attribute list

Unattached comments

type text = docstring list

val empty_text : text

val text_attr : docstring -> Parsetree.attribute

val add_text_attrs : text -> Parsetree.attributes -> Parsetree.attributes

Convert text to attributes and add them to an attribute list