Module Migrate_parsetree__Migrate_parsetree_def

type missing_feature =

| Pexp_letexception

| Ppat_open

| Pexp_unreachable

| PSig

| Pcstr_record

| Pconst_integer

| Pconst_float

| Pcl_open

| Pcty_open

| Oinherit

| Pwith_typesubst_longident

| Pwith_modsubst_longident

| Pexp_open

| Pexp_letop

| Psig_typesubst

| Psig_modsubst

| Otyp_module

| Immediate64

| Anonymous_let_module

| Anonymous_unpack

| Anonymous_module_binding

| Anonymous_module_declaration

Features which are not available in all versions of the frontend

exception Migration_error of missing_feature * Location.t

Exception thrown by migration functions when a feature is not supported.

val missing_feature_description : missing_feature -> string

missing_feature_description x is a text describing the feature x.

val missing_feature_minimal_version : missing_feature -> string

missing_feature_minimal_version x is the OCaml version where x was introduced.

val migration_error_message : missing_feature -> string

Turn a missing feature into a reasonable error message.